add function to Revision to generate 'sha1' if it does not exist
In Wikia XML dumps and probably in other dumps, <sha1></sha1> tags in revisions
are missing.

This patch coverts the 'sha1' attribute into a property that will take or set a
SHA1 checksum (e.g., if it's defined in the dump) in which case the library
behavior should be unchanged.

If, however, the SHA1 is not defined in the dump, this patch will lazily
compute the SHA1 checksum the first time the propery is accessed.

This patches uses hashlib to generate hexadecimal hash digests and does not
generate the base36 SHA1 sums that Mediawiki uses.